refactor the way back button and menu toggle are handled
There is no need to define a button with menu-toggle in each view. The toggle button will be displayed if it is a root action (all the actions in the menu are root actions). If there is a previous view that can be navigated to)

the only thing that needs to be done is to set historyRoot to true when we navigate to a rootAction. I will do something to automatize that.